src/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
pathlib
PyJWT